Remove Collection (#683)
* Remove Collection from signac API.

* Update changelog.

* Update naming/docs.

* Remove collection.py.

* Remove benchmark.py and related CI/requirements.

* Update setup.cfg
