+++ changelog.txt
@@ -34,6 +34,7 @@ Removed
  - The ``Project.config`` property is no longer mutable. Use the command line ``$ signac config`` to modify configuration (#608, #246, #244).
  - The following config related functions: ``get_config_file``, ``search_standard_dirs`` (#674).
  - ``Project`` subclasses can no longer define a ``Job`` subclass to use (#588, #693).
+ - The ``Collection`` class has been removed (#664, #667, #683).
 
 Version 1
 =========
