--- requirements/requirements-benchmark.txt
@@ -1,6 +1,0 @@
-click==8.1.3
-gitdb2==4.0.2
-GitPython==3.1.27
-numpy==1.23.3
-pandas==1.5.0; implementation_name=='cpython' --no-binary :none:
-psutil==5.9.2
